[Con] distance calculation returns float instead of numpy array element

+++ ros/src/twist_controller/dbw_common.py
@@ -30,7 +30,7 @@ def get_cross_track_error(waypoints, pose):
     return cte
 
 def distance(x1, y1, x2, y2):
-    return np.sqrt((x2-x1)**2 + (y2-y1)**2)
+    return float(np.sqrt((x2-x1)**2 + (y2-y1)**2))
 
 def get_cross_track_error_from_frenet(waypoints, pose):
